requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.20
aiohttp==3.13.2
python-dotenv==1.0.0
requests==2.31.0